Отдельный скрипт prepare_dataset для подготовки датасетов
